0.0.1

+++ contracts/lib/CheapMintNFTBytesLib.sol
@@ -102,7 +102,7 @@ library CheapMintNFTBytesLib {
     function toUint24(bytes memory _bytes, uint256 _start) internal pure returns (uint24) {
         require(_start + 3 >= _start, "toUint24_overflow");
         require(_bytes.length >= _start + 3, "toUint24_outOfBounds");
-        uint8 tempUint;
+        uint24 tempUint;
 
         assembly {
             tempUint := mload(add(add(_bytes, 0x3), _start))
